backend/requirements.txt
fastapi
uvicorn
supabase
pandas
openpyxl
psutil
numpy
faiss-cpu
qdrant-client
python-multipart
python-dotenv
httpx
PyMuPDF
ragas
langchain
langchain-community
langchain-core
langchain-ollama
langchain-groq
datasets
